final version. strsplit not working

+++ libft/ft_memcpy.c
@@ -6,20 +6,20 @@
 /*   By: tcimpean <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/10/25 22:09:57 by tcimpean          #+#    #+#             */
-/*   Updated: 2015/10/26 22:54:08 by tcimpean         ###   ########.fr       */
+/*   Updated: 2015/11/22 23:43:39 by tcimpean         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <string.h>
+#include "libft.h"
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	char	*c1;
 	char	*c2;
-	int		i;
+	size_t	i;
 
 	if (n == 0 || dst == src)
-		return(dst);
+		return (dst);
 	i = 0;
 	c1 = (char *)dst;
 	c2 = (char *)src;
@@ -28,5 +28,5 @@ void	*ft_memcpy(void *dst, const void *src, size_t n)
 		c1[i] = c2[i];
 		i++;
 	}
-	return(dst);
+	return (dst);
 }
