+++ libft/ft_strncat.c
@@ -6,16 +6,17 @@
 /*   By: tcimpean <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/08 20:45:01 by tcimpean          #+#    #+#             */
-/*   Updated: 2015/11/08 20:53:45 by tcimpean         ###   ########.fr       */
+/*   Updated: 2015/11/22 23:44:43 by tcimpean         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <string.h>
+#include "libft.h"
 
 char	*ft_strncat(char *s1, const char *s2, size_t n)
 {
-	int i;
-	int j;
+	size_t i;
+	size_t j;
+
 	i = 0;
 	while (s1[i] != '\0')
 		i++;
